OTTIMP-245: Adds collision report generator
This is helpful to know since there are scenarios where we can't update
client rate limits due to the ABA problem.

This is essentially when a user runs requests concurrently and we fail
to update the dynamodb table due to a conditional failure on another
request updating the table after we first fetch the config.